fix(server, databae): nip01.

+++ database/config.go
@@ -3,6 +3,6 @@ package database
 type Config struct {
 	URI               string
 	DBName            string `yaml:"db_name"`
-	ConnectionTimeout int32  `yaml:"connection_timeout_in_ms"`
-	QueryTimeout      int32  `yaml:"query_timeout_in_ms"`
+	ConnectionTimeout int16  `yaml:"connection_timeout_in_ms"`
+	QueryTimeout      int16  `yaml:"query_timeout_in_ms"`
 }
